Refactoring reingtest.py into modules for tutorial and documentation.
