+++ scripts/compare.sh
