+++ scripts/fig.py
