+++ scripts/fig1.py
