+++ scripts/gapspk.std
