+++ scripts/read.py
