+++ scripts/spkplt.hoc
